refactor: Add onMoveEnd event handler to BrainStormCanvas.vue
This commit adds the `onMoveEnd` event handler to the `BrainStormCanvas.vue` component. The `onMoveEnd` function is called when the user finishes moving a node on the canvas. It saves the viewport state by calling the `saveViewportState` method from the `lucidFlow` object, passing the updated `x`, `y`, and `zoom` values.

These changes improve the functionality of the BrainStormCanvas component by allowing the user to save the viewport state after moving a node.